Add missing style definitions for treeview node images
Node icons are working properly, but 5daf134 introduced optional
node image support. For this it is necessary to have some extra
style definitions.

Maybe this PR is wrong and the bootstrap-treeview.css file needs
to be included in our codebase under the lib folder and loaded
from the top level less file...