chore(regressions): Update regression tests to increase stability
There were issues with corrupted reference images, so the images were
updated to include accurate screenshots. The pdf generation was also
causing issues, so it was disabled so we can implement it
after the rest of the suite is stable.

The README had the old instructions, in addition to the new ones. I
removed the old instructions for working with the regression tests.