fix(release): patternfly-eng-release scripts now run regression tests and the automated RCUE and patternfly-org release
The patternfly-eng-release scripts also run the regression tests if 'backstop/test.js' is available. This ensures that "npm install" (if run too early) does not inadvertently install old/out-of-date dependencies prior to running the build, build test, and Shrinkwrap security vulnerability test. This also ensures that the regression tests are not run multiple times during a semantic release; once for the master branch and twice for master-dist.

Also created a new patternfly-eng-release script to kick off the automated RCUE and patternfly-org release in order to sync with PatternFly.