fix(build): Shrinkwrap should not be committed in master (#1107)
It appears that a npm-shrinkwrap.json file was inadvertently added when RCUE was introduced -- it should not be committed in master. This file is currently causing a merge conflict with master-dist; thus, preventing the semantic release from succeeding.

Note that the shrinkwrap is generated via the Travis build and committed to master-dist for bower and npm publishing.